requirements.txt
torch~=2.4.0
torchsummary~=1.5.1
numpy~=2.0.1
torchaudio~=2.4.0
matplotlib~=3.9.1
soundfile
python-dotenv
scikit-image
torchvision
tensorboard
librosa